requirements.txt
antlr4-python3-runtime
fastapi[all]
redis
spacy
